style: lint Certora specs

+++ spec/harness/CollateralTokenHarness.sol
@@ -7,7 +7,10 @@ contract CollateralTokenHarness is CollateralToken {
     ////////////////////////////////////////////////////////////////////////////
     //                         Constructors and inits                         //
     ////////////////////////////////////////////////////////////////////////////
-    constructor(address _quantConfig) public CollateralToken(_quantConfig,"","") {}
+    constructor(address _quantConfig)
+        public
+        CollateralToken(_quantConfig, "", "")
+    {}
 
     ////////////////////////////////////////////////////////////////////////////
     //                        Getters for The Internals                       //
@@ -20,40 +23,54 @@ contract CollateralTokenHarness is CollateralToken {
         return collateralTokenIds[tokenId];
     }
 
-    function getCollateralTokenInfoTokenAddress(uint collateralTokenInfoId) public view returns (address) {
+    function getCollateralTokenInfoTokenAddress(uint256 collateralTokenInfoId)
+        public
+        view
+        returns (address)
+    {
         return idToInfo[collateralTokenInfoId].qTokenAddress;
     }
 
-    function getCollateralTokenInfoTokenAsCollateral(uint collateralTokenInfoId) public view returns (address) {
+    function getCollateralTokenInfoTokenAsCollateral(
+        uint256 collateralTokenInfoId
+    ) public view returns (address) {
         return idToInfo[collateralTokenInfoId].qTokenAsCollateral;
     }
 
-    function idToInfoContainsId(uint key) public view returns (bool) {
+    function idToInfoContainsId(uint256 key) public view returns (bool) {
         if (idToInfo[key].qTokenAddress == address(0)) {
             return false;
         }
         return true;
     }
 
-    function collateralTokenIdsContainsId(uint key, uint i) public view returns (bool) {
+    function collateralTokenIdsContainsId(uint256 key, uint256 i)
+        public
+        view
+        returns (bool)
+    {
         if (collateralTokenIds[i] == key) {
             return true;
         }
         return false;
     }
 
-    function tokenSuppliesContainsCollateralToken(uint key) public view returns (bool) {
+    function tokenSuppliesContainsCollateralToken(uint256 key)
+        public
+        view
+        returns (bool)
+    {
         if (tokenSupplies[key] == uint256(0)) {
             return false;
         }
         return true;
     }
 
-    function getTokenSupplies(uint key) public view returns (uint) {
+    function getTokenSupplies(uint256 key) public view returns (uint256) {
         return tokenSupplies[key];
     }
 
-	////////////////////////////////////////////////////////////////////////////
+    ////////////////////////////////////////////////////////////////////////////
     //                       Simplifiers and override                         //
     ////////////////////////////////////////////////////////////////////////////
     /*function ...() public view returns (uint256) {
